Test reorg done, tests pass
